# openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Release Alert API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: OK
  /subscriptions:
    post:
      summary: Create email subscription for repository releases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: GitHub API rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream API failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List subscriptions
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    delete:
      summary: Deactivate subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subscription deactivated
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CreateSubscriptionRequest:
      type: object
      required: [email, repository]
      properties:
        email:
          type: string
          format: email
        repository:
          type: string
          description: owner/repo
          example: renovatebot/renovate
    Subscription:
      type: object
      required: [id, email, repository, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        repository:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string